Testing to use shallow from enzyme
While checking the HTML generated has helped me to break the code into
several components, this strategy is not great because:
 - if I decide to change the structure of the HTML the tests break,
 - it's cumbersome checking the HTML when adding new data.

Using enzyme allows me to perform "shallow" tests on the generated
output. It's a better strategy for testing components because it tests
at their level rather than the whole stack. This results in looser
coupling in the test code.