Results component created
Thought it was useful to have a separate component for deciding what
type of results we are looking at.

Because of the nature of the results from Spotify's Search API I have
decided that I will only show one type at any one time. This is because
if I get round to pagination, it only makes sense to paginate through
one type at a time.